feat(api): add playlist endpoints
#858

+++ antarest/study/storage/variantstudy/model/command/update_playlist.py
@@ -17,7 +17,7 @@ from antarest.study.storage.variantstudy.model.model import CommandDTO
 class UpdatePlaylist(ICommand):
     active: bool
     items: Optional[List[int]] = None
-    weights: Optional[Dict[int, int]] = None
+    weights: Optional[Dict[int, float]] = None
     reverse: bool = False
 
     def __init__(self, **data: Any) -> None:
